% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncov-data.R
\name{get_foreign_ncov}
\alias{get_foreign_ncov}
\title{Download ncov 2019 area and overall data from 163}
\usage{
get_foreign_ncov(country)
}
\arguments{
\item{country}{foeign country name}
}
\description{
Download ncov 2019 area and overall data from 163
}
