% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncov-data.R
\name{get_ncov2}
\alias{get_ncov2}
\title{Download ncov 2019 data from
https://github.com/BlankerL/DXY-COVID-19-Data/ (csv) or
https://github.com/yiluheihei/nCoV-2019-Data (RDS)}
\usage{
get_ncov2(latest = TRUE, method = c("ncov", "api"))
}
\arguments{
\item{latest}{logical, download the latest or all time-series ncov data,
default \code{TRUE}}

\item{method}{character}
}
\description{
Download ncov 2019 data from
https://github.com/BlankerL/DXY-COVID-19-Data/ (csv) or
https://github.com/yiluheihei/nCoV-2019-Data (RDS)
}
