% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/province_map.R
\name{plot_province_map}
\alias{plot_province_map}
\alias{plot_province_map2}
\title{Plot province maps for nCoV}
\usage{
plot_province_map(
  ncov,
  province,
  key = c("confirmedCount", "suspectedCount", "curedCount", "deadCount"),
  legend_position = c("bottomright", "topright", "bottomleft", "topleft"),
  legend_title = "Confirmed",
  color = "Reds",
  scale = c("cat", "log"),
  bins = c(0, 10, 100, 1000),
  map_title = paste0(province, "nCoV")
)

plot_province_map2(
  ncov,
  province,
  key = c("confirmedCount", "suspectedCount", "curedCount", "deadCount"),
  legend_position = c("bottomright", "topright", "bottomleft", "topleft"),
  legend_title = "Confirmed",
  color = "Reds",
  scale = c("cat", "log"),
  bins = c(0, 10, 100, 1000),
  tile_type = leaflet::addPolygons,
  map_title = paste0(province, "nCoV")
)
}
\arguments{
\item{ncov}{ncov the dataset obtained by \code{\link[ncovr]{get_ncov}}}

\item{province}{which province to plot}

\item{key}{the feature to plot}

\item{legend_position}{legend position}

\item{legend_title}{legend title}

\item{color}{color palette}

\item{scale}{category or factor}

\item{bins}{a numberic vector to cut the value to format the value to}

\item{map_title}{map title
category, only worked while \code{scale} is "cat"}

\item{tile_type}{function to define tile like amap or
\code{\link[leaflet]{addTiles}}}
}
\description{
Plot province maps for nCoV
}
