% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/province_map.R
\name{plot_province_map}
\alias{plot_province_map}
\title{Plot province maps for nCoV}
\usage{
plot_province_map(
  ncov,
  province,
  key = c("confirmedCount", "suspectedCount", "curedCount", "deadCount"),
  legend_position = c("bottomright", "topright", "bottomleft", "topleft"),
  legend_title = "Confirmed",
  color = "Reds",
  scale = c("cat", "log"),
  bins = c(0, 10, 100, 1000),
  map_title = paste0(province, "nCov map")
)
}
\arguments{
\item{ncov}{ncov the dataset obtained by \code{\link{get_ncov}}}

\item{province}{which province to plot}

\item{key}{the feature to plot}

\item{legend_position}{legend position}

\item{legend_title}{legend title}

\item{color}{color palette}

\item{scale}{category or factor}

\item{bins}{a numberic vector to cut the value to format the value to}

\item{map_title}{map title
category, only worked while \code{scale} is "cat"}
}
\description{
Plot province maps for nCoV
}
\examples{
\dontrun{
ncov <- get_ncov()
plot_province_map(ncov, "婀栧寳鐪<81>", bins = bins = c(1, 100, 500, 1000, 10000))
}
}
